pyproject.toml
[project]
name = "scheduler-email"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "django-apscheduler (>=0.7.0,<0.8.0)",
    "pytz (>=2025.2,<2026.0)",
    "python-slugify (>=8.0.4,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
